IpRetriever fonctionnel & tests ok
